+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Adding TaskUnobservedTaskExceptionIntegration to default integrations and method to remove it (#870) @FilipNemec
 - Enrich transactions with more data (#875) @Tyrrrz
 
+### Fixes
+  - Don't add version prefix in release if it's already set (#877) @Tyrrrz
+
 ## 3.0.8
 
 ### Features